End the v3 anagrams method.
**Motivation:**
There was a solution that I was missing and in the end I'm just
thinking... How didn't I thought on this. So, the thing is, loop for
every word we have available and for each sort the letters and make this
ordered letters word a key on a hash. Everytime one word with the
letters sorted has the same key it is saved.
In the end we get a hash having all the sort letters keys and each with
an array with all the words that match, all the anagrams.

Because we just go for one loop, in the end I got 15 seconds to find all
the anagrams using 355k of english words. Cool ;)